Initial upload - linux-xanmod-edge - 5.17.1
This is the initial PKGBUILD for a highly customized Xanmod Kernel with some patches on top for lower latency and a better gaming experience. I recommend to try out the ProjectC scheduler. I don't provide my Kernel config file here, as it it too special for my own needs and probably not suitable for generic use. As always, take a look at the PKGBUILD first, change the home directory and add or remove patches as you like it. This is probably the package which will see the most activity as these patches might need a rebase with newer Kernel releases.