Upload Mesa-git plus patches
Beware, this is a fast moving package, the patches might not apply cleanly anymore at a future point in time - I cannot promise to keep them up-to-date in a timely fashion but as I use it myself, I try my best. Also this is a minimal package only suitable for AMD GPU owners. You need to enable other drivers in the PKGBUILD if you are on different hardware.