Add Update script, but use it with caution
You need to make it executable first with "chmod +x". As I got tired of doing this manually I learned how to craft a basic script. Be aware that the build order is not optimal yet because of several interdependencies, hence it will only work if you already have these packages installed manually in the correct order at least once before. There could be breakage due to this. If I find the time, I might find out the correct build order and adjust the packages accordingly.