Update PKGBUILD
This is another major update, as I could not get openmp and thus polly to work with the old PKGBUILD; it turned out that there were some subtle issues in the toolchain, the wrong elf class and also a missing LLVM.gold that caused Cmake issues while trying to build openmp in the officially supported way (and only noticed today that ptr1337's toolchain contains a fix for the LLVM.gold issue that unblocks the other issue).
Notice that I adjusted the conflicts/provides section. This is not strictly needed as Arch does not have seperate lib32-subproject packages, but I tested such a openmp package to save me some time in figuring out the build issues mentioned above and it mirrors the llvm-git package.