Update PKGBUILD
I've just tested this variant to build OpenMP which was my old way of doing this (as most distributions build LLVM with OpenMP that way). The interesting part is that this is not the encouraged way to build OpenMP (https://openmp.llvm.org/SupportAndFAQ.html). On the other hand, this way should also unblock my 32-bit OpenMP/Polly experiments, so until the OpenMP devs fix the runtime route properly, simply use the old way.