Dumb down memset due to uncovered USB race condition
If memset is too fast, it turns out that it unveils a race condition in the USB enumeration/detection code. My mouse would not get detected properly and would need to be re-plugged after each re-boot to get recognized. This also happens when plugging it into a different USB slot. Hence it is not a hardware issue, but a software issue.